+++ services/web/server/src/simcore_service_webserver/catalog.py
@@ -11,6 +11,7 @@ from servicelib.aiohttp.rest_routing import iter_path_operations
 from yarl import URL
 
 from . import catalog_api_handlers, catalog_client
+from ._constants import APP_OPENAPI_SPECS_KEY, RQ_PRODUCT_KEY, X_PRODUCT_NAME_HEADER
 from .catalog_client import (
     get_services_for_user_in_product,
     is_service_responsive,
@@ -21,7 +22,6 @@ from .catalog_config import (
     KCATALOG_VERSION_PREFIX,
     assert_valid_config,
 )
-from .constants import APP_OPENAPI_SPECS_KEY, RQ_PRODUCT_KEY, X_PRODUCT_NAME_HEADER
 from .login.decorators import RQT_USERID_KEY, login_required
 from .security_decorators import permission_required
