+++ services/web/server/src/simcore_service_webserver/clusters/module_setup.py
@@ -15,8 +15,8 @@ from servicelib.aiohttp.application_setup import (
     app_module_setup,
 )
 
+from .._constants import APP_SETTINGS_KEY
 from ..application_settings import ApplicationSettings
-from ..constants import APP_SETTINGS_KEY
 from . import handlers
 
 log = logging.getLogger(__file__)
