+++ services/web/server/src/simcore_service_webserver/director/module_setup.py
@@ -12,7 +12,7 @@ from servicelib.aiohttp.rest_routing import (
     map_handlers_with_operations,
 )
 
-from ..constants import APP_OPENAPI_SPECS_KEY
+from .._constants import APP_OPENAPI_SPECS_KEY
 from .config import APP_DIRECTOR_API_KEY, assert_valid_config
 
 logger = logging.getLogger(__name__)
