+++ services/web/server/src/simcore_service_webserver/director_v2.py
@@ -8,7 +8,7 @@ from servicelib.aiohttp.rest_routing import (
 )
 
 from . import director_v2_handlers
-from .constants import APP_OPENAPI_SPECS_KEY
+from ._constants import APP_OPENAPI_SPECS_KEY
 from .director_v2_abc import set_project_run_policy
 from .director_v2_core import DefaultProjectRunPolicy, DirectorV2ApiClient, set_client
 from .director_v2_settings import CONFIG_SECTION_NAME, create_settings
