+++ services/web/server/src/simcore_service_webserver/groups.py
@@ -12,7 +12,7 @@ from servicelib.aiohttp.rest_routing import (
 )
 
 from . import groups_handlers
-from .constants import APP_OPENAPI_SPECS_KEY
+from ._constants import APP_OPENAPI_SPECS_KEY
 from .scicrunch.submodule_setup import setup_scicrunch_submodule
 
 logger = logging.getLogger(__name__)
