+++ services/web/server/src/simcore_service_webserver/login/module_setup.py
@@ -8,7 +8,7 @@ from servicelib.aiohttp.application_keys import APP_CONFIG_KEY
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 from servicelib.common_aiopg_utils import DSN
 
-from ..constants import APP_OPENAPI_SPECS_KEY, INDEX_RESOURCE_NAME
+from .._constants import APP_OPENAPI_SPECS_KEY, INDEX_RESOURCE_NAME
 from ..db_config import CONFIG_SECTION_NAME as DB_SECTION
 from .cfg import APP_LOGIN_CONFIG, cfg
 from .config import assert_valid_config, create_login_internal_config
