+++ services/web/server/src/simcore_service_webserver/projects/module_setup.py
@@ -16,7 +16,7 @@ from servicelib.aiohttp.rest_routing import (
     map_handlers_with_operations,
 )
 
-from ..constants import APP_OPENAPI_SPECS_KEY
+from .._constants import APP_OPENAPI_SPECS_KEY
 from ..resources import resources
 from . import projects_handlers, projects_nodes_handlers, projects_tags_handlers
 from .config import CONFIG_SECTION_NAME, assert_valid_config
