+++ services/web/server/tests/unit/with_dbs/04/test_change_email.py
@@ -5,7 +5,7 @@
 from aiohttp import web
 from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import LoggedUser, NewUser, parse_link
-from simcore_service_webserver.constants import INDEX_RESOURCE_NAME
+from simcore_service_webserver._constants import INDEX_RESOURCE_NAME
 from simcore_service_webserver.login.cfg import APP_LOGIN_CONFIG
 from yarl import URL
