+++ services/web/server/tests/unit/with_dbs/04/test_groups_classifiers.py
@@ -7,7 +7,7 @@ from typing import Dict
 import pytest
 import sqlalchemy as sa
 from servicelib.common_aiopg_utils import DataSourceName, create_pg_engine
-from simcore_service_webserver.constants import APP_DB_ENGINE_KEY
+from simcore_service_webserver._constants import APP_DB_ENGINE_KEY
 from simcore_service_webserver.groups_classifiers import GroupClassifierRepository
 from sqlalchemy.sql import text
